feat(3DTiles) remove todo and clean code
